removed call from backend and put it on front
realized that I would not be able to access the conversion data from the back end as it only send that info to EJS to render. Created an object with all of the currency types to populate the drop downs and moved the API call to the front so the conversion rates could be on hand for dynamic use.